+++ src/main/java/domain/piece/Soldier.java
@@ -54,7 +54,7 @@ public final class Soldier extends StaticPositionedPiece {
     }
 
     @Override
-    public int getScore() {
+    public double getScore() {
         return 2;
     }
