+++ src/test/java/domain/piece/GuardTest.java
@@ -50,10 +50,10 @@ class GuardTest {
     void 본인의_점수를_반환한다() {
         // given
         Guard guard = new Guard(Side.HAN);
-        int expected = 3;
+        double expected = 3;
 
         // when
-        int actual = guard.getScore();
+        double actual = guard.getScore();
 
         // then
         assertThat(actual).isEqualTo(expected);
